requirements.txt
beautifulsoup4==4.11.1
langchain==0.3.5
langchain_chroma==0.1.4
langchain_community==0.3.3
langchain_core==0.3.13
langchain_openai==0.2.4
langchain_text_splitters==0.3.1
openai==1.53.0
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.39.0
